removes some eslint issues, fixes prettier and eslint discrepencises
