+++ .eslintrc.json
@@ -1,10 +1,11 @@
 {
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
+  "plugins": ["@typescript-eslint", "prettier"],
   "extends": [
     "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:@typescript-eslint/recommended-requiring-type-checking"
+    "plugin:@typescript-eslint/recommended"
+    //"plugin:@typescript-eslint/recommended-requiring-type-checking",
+    //"plugin:prettier/recommended"
   ],
   "parserOptions": {
     "ecmaVersion": 2020,
@@ -16,6 +17,7 @@
     "project": ["./tsconfig.json"]
   },
   "rules": {
+    "quotes": ["error", "single", { "avoidEscape": true }]
     // Add your custom rules here
   }
 }
